requirements.txt
requests==2.28.2
Pillow==9.5.0
imagehash==4.3.1
PySocks==1.7.1
urllib3==1.26.15
# python-tumblr==1.1.4  # 代替手段を検討中
tumblr-api==0.1.4  # 代替パッケージ
numpy==1.24.3
opencv-python==4.7.0.72
scikit-image==0.20.0
tensorflow==2.12.0
keras==2.12.0
# transformers==4.29.2

# ロギングとデバッグ用
python-dotenv==1.0.0
pyyaml==6.0

# テスト用
pytest==7.3.1
pytest-cov==4.0.0

# 開発ツール
black==23.3.0
flake8==6.0.0
mypy==1.3.0
